Extract common code of callRustInSameThreadSync
