+++ zaplib/web/zaplib_worker_runtime.ts
@@ -10,12 +10,12 @@
 // Currently this is only supported in WebAssembly, not when using CEF.
 
 import {
+  callRustInSameThreadSyncImpl,
   createErrorCheckers,
   createWasmBuffer,
   getWasmEnv,
   getZapParamType,
   initThreadLocalStorageAndStackOtherWorkers,
-  makeZerdeBuilder,
   Rpc,
   transformParamsFromRustImpl,
 } from "common";
@@ -25,7 +25,6 @@ import {
   CallRustInSameThreadSync,
   PostMessageTypedArray,
   WasmExports,
-  ZapParamType,
   ZapArray,
   RustZapParam,
   MutableBufferData,
@@ -40,7 +39,6 @@ import {
   ZapBuffer,
   checkValidZapArray,
 } from "zap_buffer";
-import { ZerdeParser } from "zerde";
 
 let rpc: Rpc<WebWorkerRpc>;
 let wasmExports: WasmExports;
@@ -177,64 +175,19 @@ export const callRust: CallRust = async (name, params = []) => {
   );
 };
 
-// TODO(JP): Some of this code is duplicated with callRust/call_js; see if we can reuse some.
 export const callRustInSameThreadSync: CallRustInSameThreadSync = (
   name,
   params = []
-) => {
-  checkWasm();
-
-  const zerdeBuilder = makeZerdeBuilder(wasmMemory, wasmExports);
-  zerdeBuilder.sendString(name);
-  zerdeBuilder.sendU32(params.length);
-  for (const param of params) {
-    if (typeof param === "string") {
-      zerdeBuilder.sendU32(ZapParamType.String);
-      zerdeBuilder.sendString(param);
-    } else {
-      if (param.buffer instanceof ZapBuffer) {
-        checkValidZapArray(param);
-        if (param.buffer.__zaplibBufferData.readonly) {
-          zerdeBuilder.sendU32(getZapParamType(param, true));
-
-          const arcPtr = param.buffer.__zaplibBufferData.arcPtr;
-
-          // ZapParam parsing code will construct an Arc without incrementing
-          // the count, so we do it here ahead of time.
-          wasmExports.incrementArc(BigInt(arcPtr));
-          zerdeBuilder.sendU32(arcPtr);
-        } else {
-          // TODO(Paras): User should not be able to access the buffer after
-          // passing it to Rust here
-          unregisterMutableBuffer(param.buffer);
-          zerdeBuilder.sendU32(getZapParamType(param, false));
-          zerdeBuilder.sendU32(param.buffer.__zaplibBufferData.bufferPtr);
-          zerdeBuilder.sendU32(param.buffer.__zaplibBufferData.bufferLen);
-          zerdeBuilder.sendU32(param.buffer.__zaplibBufferData.bufferCap);
-        }
-      } else {
-        console.warn(
-          "Consider passing Uint8Arrays backed by ZapBuffer to prevent copying data"
-        );
-
-        const vecLen = param.byteLength;
-        const vecPtr = createWasmBuffer(wasmMemory, wasmExports, param);
-        zerdeBuilder.sendU32(getZapParamType(param, false));
-        zerdeBuilder.sendU32(vecPtr);
-        zerdeBuilder.sendU32(vecLen);
-        zerdeBuilder.sendU32(vecLen);
-      }
-    }
-  }
-  const returnPtr = wasmExports.callRustInSameThreadSync(
+) =>
+  callRustInSameThreadSyncImpl({
+    name,
+    params,
+    checkWasm,
+    wasmMemory,
+    wasmExports,
     wasmAppPtr,
-    BigInt(zerdeBuilder.getData().byteOffset)
-  );
-
-  const zerdeParser = new ZerdeParser(wasmMemory, Number(returnPtr));
-  const returnParams = zerdeParser.parseZapParams();
-  return transformParamsFromRust(returnParams);
-};
+    transformParamsFromRust,
+  });
 
 // TODO(JP): See comment at CreateBufferWorkerSync type.
 export const createMutableBuffer: CreateBufferWorkerSync = (data) => {
